+++ my_flutter/lib/src/resources/repositories/ticket_repository.dart
@@ -1,4 +1,3 @@
-import 'dart:ffi';
 
 import 'package:my_flutter/src/models/lists/tickets_list.dart';
 import 'package:my_flutter/src/models/ticket_model.dart';
@@ -12,9 +11,9 @@ class Repository {
   Future<TicketModel> buyTicket(TicketModel ticket) => ticketsApiProvider.buyTicket(ticket);
   Future<TicketModel> mailTicket(TicketModel ticket) => ticketsApiProvider.mailTicket(ticket);
   Future<TicketModel> exchangeTicket(RequestModel ticket) => ticketsApiProvider.exchangeTicket(ticket);
-  Future<TicketsList> fetchTicketsByUserId(Long userId) => ticketsApiProvider.fetchTicketsByUserId(userId);
-  Future<TicketModel> fetchTicket(Long id) => ticketsApiProvider.fetchTicket(id);
-  Future<TicketModel> deleteTicket(Long id) => ticketsApiProvider.deleteTicket(id);
+  Future<TicketsList> fetchTicketsByUserId(int userId) => ticketsApiProvider.fetchTicketsByUserId(userId);
+  Future<TicketModel> fetchTicket(int id) => ticketsApiProvider.fetchTicket(id);
+  Future<TicketModel> deleteTicket(int id) => ticketsApiProvider.deleteTicket(id);
   Future<TicketModel> addTicket(TicketModel ticket) => ticketsApiProvider.addTicket(ticket);
   Future<TicketModel> updateTicket(TicketModel ticket) => ticketsApiProvider.updateTicket(ticket);
